# bin/Release/net7.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Aplicada2Api, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null",
    "version": "1.0"
  },
  "paths": {
    "/api/Citas": {
      "get": {
        "tags": [
          "Citas"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Citas"
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Citas"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Citas"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Citas"
            }
          }
        }
      }
    },
    "/api/Citas/{id}": {
      "get": {
        "tags": [
          "Citas"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Citas"
            }
          }
        }
      },
      "put": {
        "tags": [
          "Citas"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Citas"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Citas"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Clientes": {
      "get": {
        "tags": [
          "Clientes"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Clientes"
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Clientes"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Clientes"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Clientes"
            }
          }
        }
      }
    },
    "/api/Clientes/{id}": {
      "get": {
        "tags": [
          "Clientes"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Clientes"
            }
          }
        }
      },
      "put": {
        "tags": [
          "Clientes"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Clientes"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Clientes"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Mecanicos": {
      "get": {
        "tags": [
          "Mecanicos"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Mecanicos"
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Mecanicos"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Mecanicos"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Mecanicos"
            }
          }
        }
      }
    },
    "/api/Mecanicos/{id}": {
      "get": {
        "tags": [
          "Mecanicos"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Mecanicos"
            }
          }
        }
      },
      "put": {
        "tags": [
          "Mecanicos"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Mecanicos"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Mecanicos"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  },
  "definitions": {
    "Citas": {
      "type": "object",
      "properties": {
        "citaId": {
          "format": "int32",
          "type": "integer"
        },
        "concepto": {
          "type": "string"
        },
        "fecha": {
          "format": "date-time",
          "type": "string"
        },
        "clienteId": {
          "format": "int32",
          "type": "integer"
        },
        "mecanicoId": {
          "format": "int32",
          "type": "integer"
        },
        "clientes": {
          "$ref": "#/definitions/Clientes"
        },
        "mecanicos": {
          "$ref": "#/definitions/Mecanicos"
        }
      },
      "additionalProperties": false
    },
    "Clientes": {
      "type": "object",
      "properties": {
        "clienteId": {
          "format": "int32",
          "type": "integer"
        },
        "nombres": {
          "type": "string"
        },
        "telefono": {
          "type": "string"
        },
        "direccion": {
          "type": "string"
        },
        "vehiculoId": {
          "format": "int32",
          "type": "integer"
        },
        "vehiculos": {
          "$ref": "#/definitions/Vehiculos"
        }
      },
      "additionalProperties": false
    },
    "Mecanicos": {
      "type": "object",
      "properties": {
        "mecanicoId": {
          "format": "int32",
          "type": "integer"
        },
        "nombres": {
          "type": "string"
        },
        "area": {
          "type": "string"
        },
        "telefono": {
          "type": "string"
        },
        "disponible": {
          "format": "int32",
          "type": "integer"
        }
      },
      "additionalProperties": false
    },
...
    "Vehiculos": {
      "type": "object",
      "properties": {
        "vehiculoId": {
          "format": "int32",
          "type": "integer"
        },
        "marca": {
          "type": "string"
        },
        "modelo": {
          "type": "string"
        },
        "year": {
          "format": "date-time",
          "type": "string"
        }
      },
      "additionalProperties": false
    }
  }
}